Crate cumulus_pallet_dmp_queue

This pallet used to implement a message queue for downward messages from the relay-chain.

It is now deprecated and has been refactored to simply drain any remaining messages into something implementing HandleMessage. It proceeds in the state of MigrationState one by one by their listing in the source code. The pallet can be removed from the runtime once Completed was emitted.

Re-exports

pub use weights::WeightInfo;

pub use pallet::*;Deprecated

Modules

palletDeprecated

The pallet module in each FRAME pallet hosts the most important items needed to construct this pallet.

weights

Autogenerated weights for cumulus_pallet_dmp_queue

Type Aliases

MaxDmpMessageLenOf

The maximal length of a DMP message.