Crate cumulus_pallet_dmp_queue

Pallet implementing a message queue for downward messages from the relay-chain. Executes downward messages if there is enough weight available and schedules the rest for later execution (by on_idle or another handle_dmp_messages call). Individual overweight messages are scheduled into a separate queue that is only serviced by explicit extrinsic calls.

Re-exports

pub use pallet::*;

Modules

pallet

The module that hosts all the FRAME types needed to add this pallet to a runtime.

Structs

ConfigData

PageIndexData

Information concerning our message pages.

Type Definitions

MessageId

Simple type used to identify messages for the purpose of reporting events. Secure if and only if the message content is unique.

OverweightIndex

Index used to identify overweight messages.

PageCounter

Index used to identify normal pages.